+++ tests/integration/reusable_values.py
@@ -1,11 +1,6 @@
 import asyncio
 
-from tests.integration.it_utils import (
-    ASYNC_JSON_RPC_TESTNET_CLIENT,
-    fund_wallet,
-    sign_and_reliable_submission_async,
-)
-from xrpl.asyncio.wallet import generate_faucet_wallet
+from tests.integration.it_utils import fund_wallet, sign_and_reliable_submission_async
 from xrpl.models.amounts import IssuedCurrencyAmount
 from xrpl.models.transactions import OfferCreate, PaymentChannelCreate
 from xrpl.wallet import Wallet
@@ -19,9 +14,6 @@ async def _set_up_reusable_values():
     await fund_wallet(WALLET)
     DESTINATION = Wallet.create()
     await fund_wallet(DESTINATION)
-
-    TESTNET_WALLET = await generate_faucet_wallet(ASYNC_JSON_RPC_TESTNET_CLIENT)
-    TESTNET_DESTINATION = await generate_faucet_wallet(ASYNC_JSON_RPC_TESTNET_CLIENT)
 
     OFFER = await sign_and_reliable_submission_async(
         OfferCreate(
@@ -54,8 +46,6 @@ async def _set_up_reusable_values():
     return (
         WALLET,
         DESTINATION,
-        TESTNET_WALLET,
-        TESTNET_DESTINATION,
         OFFER,
         PAYMENT_CHANNEL,
     )
@@ -64,8 +54,6 @@ async def _set_up_reusable_values():
 (
     WALLET,
     DESTINATION,
-    TESTNET_WALLET,
-    TESTNET_DESTINATION,
     OFFER,
     PAYMENT_CHANNEL,
 ) = asyncio.run(_set_up_reusable_values())