+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new ways of wallet generation - from public/private key, from entropy, from secret
 - Changed wallet constructor arguments to optional for easier inferencing
 
+### Changed:
+- `check_fee` now has a higher limit that is less likely to be hit
+- When connected to nft devnet or hooks v2 testnet generate_faucet_wallet now defaults to using the faucet instead of requiring specification
+
 ## [1.7.0] - 2022-10-12
 ### Added:
 - Support for ExpandedSignerList amendment that expands the maximum signer list to 32 entries